+++ source/extensions/omni.isaac.core/omni/isaac/core/tests/__init__.py
@@ -16,5 +16,7 @@ from .utils.test_rotations import *
 from .utils.test_stage import *
 from .utils.test_bounds import *
 from .utils.test_viewports import *
+from .utils.test_physics import *
+from .utils.test_prims import *
 from .world.test_world import *
 from .simulation_context.test_simulation_context import *